+++ src/Stratis.Bitcoin/Consensus/ProvenHeadersConsensusManagerBehavior.cs
@@ -3,35 +3,37 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Stratis.Bitcoin.Connection;
-using Stratis.Bitcoin.Consensus;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.P2P.Protocol;
 using Stratis.Bitcoin.P2P.Protocol.Payloads;
 using Stratis.Bitcoin.Utilities;
 
-namespace Stratis.Bitcoin.Features.Consensus.Behaviors
+namespace Stratis.Bitcoin.Consensus
 {
-    /// <summary>Behavior that takes care of proven headers protocol. It also keeps the notion of peer's consensus tip.</summary>
+    /// <summary>
+    /// Behavior that takes care of proven headers protocol. It also keeps the notion of peer's consensus tip.
+    /// </summary>
     public class ProvenHeadersConsensusManagerBehavior : ConsensusManagerBehavior
     {
         private readonly IInitialBlockDownloadState initialBlockDownloadState;
         private readonly IConsensusManager consensusManager;
         private readonly IPeerBanning peerBanning;
         private readonly ILoggerFactory loggerFactory;
-        private readonly PosConsensusFactory consensusFactory;
+        private readonly Network network;
 
         /// <summary>Instance logger.</summary>
         private readonly ILogger logger;
 
-        public ProvenHeadersConsensusManagerBehavior(ConcurrentChain chain, IInitialBlockDownloadState initialBlockDownloadState, IConsensusManager consensusManager, IPeerBanning peerBanning, ILoggerFactory loggerFactory) : base(chain, initialBlockDownloadState, consensusManager, peerBanning, loggerFactory)
+        public ProvenHeadersConsensusManagerBehavior(ConcurrentChain chain, IInitialBlockDownloadState initialBlockDownloadState,
+            IConsensusManager consensusManager, IPeerBanning peerBanning, ILoggerFactory loggerFactory, Network network) : base(chain, initialBlockDownloadState, consensusManager, peerBanning, loggerFactory)
         {
             this.chain = chain;
             this.initialBlockDownloadState = initialBlockDownloadState;
             this.consensusManager = consensusManager;
             this.peerBanning = peerBanning;
+            this.network = network;
             this.loggerFactory = loggerFactory;
-            this.consensusFactory = new PosConsensusFactory();
             this.logger = loggerFactory.CreateLogger(this.GetType().FullName, $"[{this.GetHashCode():x}] ");
         }
 
@@ -78,7 +80,7 @@ namespace Stratis.Bitcoin.Features.Consensus.Behaviors
             foreach (ChainedHeader header in this.chain.EnumerateToTip(fork).Skip(1))
             {
                 var posBock = new PosBlock(header.Header);
-                ProvenBlockHeader provenBlockHeader = this.consensusFactory.CreateProvenBlockHeader(posBock);
+                ProvenBlockHeader provenBlockHeader = ((PosConsensusFactory)this.network.Consensus.ConsensusFactory).CreateProvenBlockHeader(posBock);
                 lastHeader = header;
                 headers.Headers.Add(provenBlockHeader);
 
@@ -97,7 +99,8 @@ namespace Stratis.Bitcoin.Features.Consensus.Behaviors
         /// <inheritdoc />
         public override object Clone()
         {
-            return new ProvenHeadersConsensusManagerBehavior(this.chain, this.initialBlockDownloadState, this.consensusManager, this.peerBanning, this.loggerFactory);
+            return new ProvenHeadersConsensusManagerBehavior(
+                this.chain, this.initialBlockDownloadState, this.consensusManager, this.peerBanning, this.loggerFactory, this.network);
         }
     }
 }
